src/requirements.txt
# Function dependencies, for example:
# package>=version

numpy 
google-cloud-bigquery